falta dibujarlo

+++ practica 2/src/Modelo/Pair.java
@@ -2,9 +2,9 @@ package Modelo;
 
 public class Pair {
 	private int i;
-	private int j;
+	private Double j;
 
-	public Pair(int i, int j) {
+	public Pair(int i, Double j) {
 		this.i = i;
 		this.j = j;
 	}
@@ -13,7 +13,7 @@ public class Pair {
 		return i;
 	}
 
-	public int getJ() {
+	public Double getJ() {
 		return j;
 	}
 
@@ -21,7 +21,7 @@ public class Pair {
 		this.i = i;
 	}
 
-	public void setJ(int j) {
+	public void setJ(Double j) {
 		this.j = j;
 	}
